requirements.txt
streamlit
numpy
pandas
Pillow
rdkit
scikit-learn==0.22.2.post1